fixed perms

+++ app/Http/Controllers/Joystick/CategoryController.php
@@ -38,6 +38,10 @@ class CategoryController extends Controller
 
     public function store(Request $request)
     {
+        if (!\Auth::user()->can('create-role')) {
+            return redirect()->back()->with('status', 'Ваши права ограничены!');
+        }
+
         $this->validate($request, [
             'title' => 'required|min:2|max:80',
         ]);
@@ -77,6 +81,10 @@ class CategoryController extends Controller
 
     public function update(Request $request, $lang, $id)
     {
+        if (!\Auth::user()->can('edit-role')) {
+            return redirect()->back()->with('status', 'Ваши права ограничены!');
+        }
+
         $this->validate($request, [
             'title' => 'required|min:2|max:80',
         ]);
@@ -108,6 +116,10 @@ class CategoryController extends Controller
 
     public function destroy($lang, $id)
     {
+        if (!\Auth::user()->can('delete-role')) {
+            return redirect()->back()->with('status', 'Ваши права ограничены!');
+        }
+
         $category = Category::find($id);
         $category->delete();
